pyproject.toml
[project]
name = "lead-scraper"
version = "0.1.0"
description = "CRM lead scraper microservice using Scrapling + FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "scrapling[fetchers]>=0.2.9",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
